Move KeyboardManager types to class file
